Refactor factories from tests to within the app. This is so we can generate dummy data.
