+++ tests/integration/test_home_page.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.factories import ConsultationFactory
+from consultation_analyser.consultations.factories import ConsultationFactory
 
 
 @pytest.mark.django_db
