+++ tests/integration/test_responses_pages.py
@@ -2,7 +2,7 @@ import html
 
 import pytest
 
-from tests.factories import ConsultationFactory
+from consultation_analyser.consultations.factories import ConsultationFactory
 
 
 @pytest.mark.django_db
